Fix Balance Precision (#23)
number valued balance fields are essentially always wrong and shouldn't exist. Luckily Zerion returns a structured data type with the full precision already available; that is

"quantity": {
  "int": "8431170095597067",
  "decimals": 18,
  "float": 0.0084311700955971,
  "numeric": "0.008431170095597067"
}

So, this PR just changes the number type to string and uses the "numeric" value instead of the "float" value.

We made a compromise so not to break the interface. This involves using the "numeric" field value, but casting back to Number. If necessary we can take this one step further, by returning the string field.